Layout as vue file, re-arrange files
