+++ src/content/netgame.js
@@ -1,5 +1,9 @@
 console.log('Thank you for choosing KC3改 on Firefox!')
 
+// Apply global styles
 import './assets/scss/clean.netgame.scss'
 import './assets/scss/layout.scss'
-import './lib/scripts/layout'
+
+// Create viewport for KC3 layout and start layout feature
+import './scripts/create-viewport'
+import './scripts/start-layout'
