+++ src/content/scripts/create-viewport.js
@@ -1,0 +1,6 @@
+module.exports = (function() {
+  let viewport = document.createElement('div')
+  viewport.id = 'kc3viewport'
+  viewport.style = 'width:100%; background:#ccffcc;'
+  document.querySelector('body').appendChild(viewport)
+})()
