+++ src/content/scripts/start-layout.js
@@ -1,0 +1,5 @@
+import Vue from 'vue'
+import LayoutVue from '@widgets/layout.vue'
+module.exports = (function() {
+  new Vue(LayoutVue)
+})()
